added Employee Tests and Changed float type to double

+++ backend/src/main/java/com/shifts/backend/model/TimeBlock.java
@@ -2,8 +2,8 @@ package com.shifts.backend.model;
 
 public class TimeBlock {
     private int timeBlockId;
-    private float startTime;
-    private float endTime;
+    private double startTime;
+    private double endTime;
     private int shiftsRequired;
     private Enum<weekDayEnum> weekDayEnum;
 
@@ -34,16 +34,16 @@ public class TimeBlock {
     public void setTimeBlockId(int timeBlockId) {
         this.timeBlockId = timeBlockId;
     }
-    public float getStartTime() {
+    public double getStartTime() {
         return startTime;
     }
-    public void setStartTime(float startTime) {
+    public void setStartTime(double startTime) {
         this.startTime = startTime;
     }
-    public float getEndTime() {
+    public double getEndTime() {
         return endTime;
     }
-    public void setEndTime(float endTime) {
+    public void setEndTime(double endTime) {
         this.endTime = endTime;
     }
     public Enum<weekDayEnum> getWeekDayEnum() {
